Fixed compiler errors in Roslyn.Diagnostics.Analyzers.UnitTests (#1018)
Overview of typical issues in unittest code fragments:
* Duplicate parameter names
* Non-void method must return a value or throw
* Mismatch between VB array size and number of elements

+++ src/Roslyn.Diagnostics.Analyzers/UnitTests/SpecializedEnumerableCreationAnalyzerTests.cs
@@ -208,7 +208,7 @@ Imports System.Collections.Generic
 
 Class C
     Function M1() As IEnumerable(Of Integer)
-        Return New Integer(1) {1}
+        Return New Integer(0) {1}
     End Function
     Function M2() As IEnumerable(Of Integer)
         Return {1}
@@ -267,7 +267,7 @@ Imports System.Collections.Generic
 
 Class C
     Function M1() As IEnumerable(Of Integer)
-        Return New Integer(2) {1, 2}
+        Return New Integer(1) {1, 2}
     End Function
     Function M2() As IEnumerable(Of Integer)
         Return {1, 2}
